# Conflicts:
#	features/merchant/shop_page/src/main/java/com/tokopedia/shop/product/view/viewholder/ShopProductItemBigGridViewHolder.kt
#	features/merchant/shop_page/src/main/java/com/tokopedia/shop/product/view/viewholder/ShopProductItemListViewHolder.kt

+++ features/top_ads/top_ads_edit/src/main/java/com/tokopedia/topads/edit/view/fragment/edit/EditGroupAdFragment.kt
@@ -114,7 +114,7 @@ class EditGroupAdFragment : BaseDaggerFragment() {
         if (priceDaily != 0.0F) {
             toggle?.isChecked = true
             daily_budget?.visible()
-            setCurrentDailyBudget((priceDaily).toString())
+            setCurrentDailyBudget((priceDaily).toInt().toString())
         } else {
             daily_budget?.gone()
         }
